Refacto modal + create request add project
